+++ cmd/server/external/database/sqlPatient.go
@@ -5,11 +5,11 @@ import (
 	"parcial/internal/domain"
 )
 
-type SqlStore struct {
+type SqlStorePatient struct {
 	DB *sql.DB
 }
 
-func (s *SqlStore) GetPatientById(id int) (*domain.Patient, error) {
+func (s *SqlStorePatient) GetPatientById(id int) (*domain.Patient, error) {
 	var patient domain.Patient
 	query := "SELECT * FROM patient WHERE ID = ?;"
 	row := s.DB.QueryRow(query, id)
@@ -19,7 +19,7 @@ func (s *SqlStore) GetPatientById(id int) (*domain.Patient, error) {
 	}
 	return &patient, nil
 }
-func (s *SqlStore) GetPatientByDni(dni string) (*domain.Patient, error) {
+func (s *SqlStorePatient) GetPatientByDni(dni string) (*domain.Patient, error) {
 	var patient domain.Patient
 	query := "SELECT * FROM patient WHERE DNI = ?;"
 	row := s.DB.QueryRow(query, dni)
@@ -31,7 +31,7 @@ func (s *SqlStore) GetPatientByDni(dni string) (*domain.Patient, error) {
 	return &patient, nil
 }
 
-func (s *SqlStore) CreatePatient(patient domain.Patient) (*domain.Patient, error) {
+func (s *SqlStorePatient) CreatePatient(patient domain.Patient) (*domain.Patient, error) {
 	query := "INSERT INTO patient (Name, LastName, Address, DNI, DischargeDate) VALUES (?, ?, ?, ?, ?)"
 	stmt, err := s.DB.Prepare(query)
 	if err != nil {
@@ -51,7 +51,7 @@ func (s *SqlStore) CreatePatient(patient domain.Patient) (*domain.Patient, error
 	return &patient, nil
 }
 
-func (s *SqlStore) DeletePatient(id int) error {
+func (s *SqlStorePatient) DeletePatient(id int) error {
 	query := "DELETE FROM patient WHERE ID = ?"
 	_, err := s.DB.Exec(query, id)
 	if err != nil {
@@ -60,7 +60,7 @@ func (s *SqlStore) DeletePatient(id int) error {
 	return nil
 }
 
-func (s *SqlStore) UpdatePatient(id int, patient domain.Patient) (*domain.Patient, error) {
+func (s *SqlStorePatient) UpdatePatient(id int, patient domain.Patient) (*domain.Patient, error) {
 	updateQuery := "UPDATE patient SET Name = ?, LastName = ?, Address = ?, DischargeDate = ? WHERE ID = ?"
 	_, err := s.DB.Exec(updateQuery, patient.Name, patient.LastName, patient.Address, patient.DischargeDate, id)
 	if err != nil {
